Support opening issues in other repos
Extends the `open "..."` command to support `open "..." in {repo}`,
where repo is one of "prefect", "ui", "server" or "core" (alias for
"prefect"). The repo name is case agnostic.

Also refactors the parsing and formatting code out a bit to make it more
testable, and adds a test. Also makes the issue creating code more
whitespace agnostic.